+++ lib/pagy/extras/support.rb
@@ -14,25 +14,25 @@ class Pagy # :nodoc:
       pagy_url_for(pagy, pagy.next, absolute:) if pagy.next
     end
 
-    # Return the HTML string for the enabled/disabled previous page link
-    def pagy_prev_html(pagy, text: pagy_t('pagy.prev'), aria_label: pagy_t('pagy.aria_label.prev'))
+    # Return the enabled/disabled previous page anchor tag
+    def pagy_prev_a(pagy, text: pagy_t('pagy.prev'), aria_label: pagy_t('pagy.aria_label.prev'))
       a = pagy_anchor(pagy)
       prev_html(pagy, a, text:, aria_label:)
     end
 
-    # Return the HTML string for the enabled/disabled next page link
-    def pagy_next_html(pagy, text: pagy_t('pagy.next'), aria_label: pagy_t('pagy.aria_label.prev'))
+    # Return the enabled/disabled next page anchor tag
+    def pagy_next_a(pagy, text: pagy_t('pagy.next'), aria_label: pagy_t('pagy.aria_label.prev'))
       a = pagy_anchor(pagy)
       next_html(pagy, a, text:, aria_label:)
     end
 
-    # Conditionally return the HTML link tag string for the previous page
-    def pagy_prev_link_tag(pagy, absolute: false)
+    # Conditionally return the previous page link tag
+    def pagy_prev_link(pagy, absolute: false)
       %(<link href="#{pagy_url_for(pagy, pagy.prev, absolute:)}"/>) if pagy.prev
     end
 
-    # Conditionally return the HTML link tag string for the next page
-    def pagy_next_link_tag(pagy, absolute: false)
+    # Conditionally return the next page link tag
+    def pagy_next_link(pagy, absolute: false)
       %(<link href="#{pagy_url_for(pagy, pagy.next, absolute:)}"/>) if pagy.next
     end
   end